Almost everything is fixed
Responsive content and all known issues is fixed

+++ lib/data/models/availableBalance_model.dart
@@ -1,7 +1,7 @@
 class AvailableBalanceModel {
-  int? mutualFunds;
-  int? stocks;
-  int? fixedDeposits;
+  double? mutualFunds;
+  double? stocks;
+  double? fixedDeposits;
 
   AvailableBalanceModel({this.mutualFunds, this.stocks, this.fixedDeposits});
